Encode value semantics of Channel<T>
The extension axiom now requires the guards of matching sends and
receives to be true. In addition, every matching pair of send and
receive event implies that their terms are equal. This new feature
permits us to detect communication deadlocks even when the control-
flow in the program under scrutiny depends on sends/receives.